+++ examples/pixart_sigma/pixart/lora-finetuning/inference_pixart_lora.py
@@ -12,7 +12,7 @@ import argparse
 import mindspore as ms
 
 from mindone.diffusers import PixArtAlphaPipeline, Transformer2DModel
-from mindone.diffusers._peft import PeftModel
+from mindone.peft import PeftModel
 
 
 def parse_args():
