+++ examples/pixart_sigma/pixart/lora-finetuning/train_pixart_lora.py
@@ -40,9 +40,6 @@ from mindone.diffusers import (
     StableDiffusionPipeline,
     Transformer2DModel,
 )
-from mindone.diffusers._peft import LoraConfig, get_peft_model
-from mindone.diffusers._peft.tuners.tuners_utils import BaseTunerLayer
-from mindone.diffusers._peft.utils import get_peft_model_state_dict
 from mindone.diffusers.optimization import get_scheduler
 from mindone.diffusers.training_utils import (
     AttrJitWrapper,
@@ -54,6 +51,9 @@ from mindone.diffusers.training_utils import (
     set_seed,
 )
 from mindone.diffusers.utils import convert_state_dict_to_diffusers
+from mindone.peft import LoraConfig, get_peft_model
+from mindone.peft.tuners.tuners_utils import BaseTunerLayer
+from mindone.peft.utils import get_peft_model_state_dict
 
 logger = logging.getLogger(__name__)
