+++ mindone/diffusers/pipelines/pipeline_loading_utils.py
@@ -323,7 +323,7 @@ def warn_deprecated_model_variant(pretrained_model_name_or_path, token, variant,
 
 def _unwrap_model(model):
     """Unwraps a model."""
-    from mindone.diffusers._peft import PeftModel
+    from mindone.peft import PeftModel
 
     if isinstance(model, PeftModel):
         model = model.base_model.model
