+++ mindone/diffusers/utils/testing_utils.py
@@ -219,7 +219,7 @@ def require_peft_version_greater(peft_version):
     """
 
     def decorator(test_case):
-        from mindone.diffusers._peft import __version__ as _mindone_diffusers_peft_version
+        from mindone.peft import __version__ as _mindone_diffusers_peft_version
 
         correct_peft_version = version.parse(_mindone_diffusers_peft_version) > version.parse(peft_version)
         return pytest.mark.skipif(
