+++ mindone/peft/tuners/lora/model.py
@@ -60,7 +60,7 @@ class LoraModel(BaseTuner):
 
         ```py
         >>> from transformers import AutoModelForSeq2SeqLM
-        >>> from mindone.diffusers._peft import LoraModel, LoraConfig
+        >>> from mindone.peft import LoraModel, LoraConfig
 
         >>> config = LoraConfig(
         ...     task_type="SEQ_2_SEQ_LM",
@@ -76,7 +76,7 @@ class LoraModel(BaseTuner):
 
         ```py
         >>> import transformers
-        >>> from mindone.diffusers._peft import LoraConfig, PeftModel, get_peft_model
+        >>> from mindone.peft import LoraConfig, PeftModel, get_peft_model
 
         >>> target_modules = ["q_proj", "k_proj", "v_proj", "out_proj", "fc_in", "fc_out", "wte"]
         >>> config = LoraConfig(
@@ -591,7 +591,7 @@ class LoraModel(BaseTuner):
 
         ```py
         >>> from transformers import AutoModelForCausalLM
-        >>> from mindone.diffusers._peft import PeftModel
+        >>> from mindone.peft import PeftModel
 
         >>> base_model = AutoModelForCausalLM.from_pretrained("tiiuae/falcon-40b")
         >>> peft_model_id = "smangrul/falcon-40B-int4-peft-lora-sfttrainer-sample"
